Simplify hint modal using ActiveModal logic
